api 엔드포인트 토큰 검증 단계 추가

+++ api-server/src/utils/fileUtil.ts
@@ -19,7 +19,8 @@ export const uploadImage = async (file: Express.Multer.File, fulPath: string, di
 
 export const deleteImage = async (delFilePath: string) => {
     try {
-
+        if (!delFilePath) return true;
+        
         const delFullPath:string = path.resolve(delFilePath);
         await fs.unlink(delFullPath);
